fix: show forwarded client ip in live logs
